Moves assignment and configuration methods into a new class
